Fix broken build from distribution
Reverts the effect of commit 253, and partially 245. Removing Dict files
from the list of sources prevents the dictionaries being built. Retain
the .cxx Dict files in the source list, but the .h files are removed
(not needed).

Instead add a dist-hook to Makefile.am to remove Dict files before
distribution is created.